# Add support for `txt` file extensions and display hash for non text/png files