--- a/tests/testthat/test.R
+++ b/tests/testthat/test.R
@@ -1,5 +1,0 @@
-test_that("shinytest works", {
-
-  expect_true(TRUE)
-
-})